Adds publication button and isolates post update for reuse
